Migrate to null-safety
I believe this is correct.  Almost everything seems to be handleable by "late" since none of these sites actually want a null value, just wanted to initialize them after the constructor.

This found a lack of null (or general error) handling in Maze.parse which I noted with a comment.