Api refactor (#11)
Broadly, this alternate API gets rid of any assumptions about a finite
or enumerable grid, and thus allows users to plan a path across an
infinite grid or theoretical grid, such as those representing player
moves in a game.

The new API looks like this: 
- `AStarState` objects to describe the state, how to `expand()`, how to
calculate the `heuristic()`, and `isGoal`
- `AStarNode` objects wrap the state, caches heuristics and hashes, then
tracks depth and parents
- Each state has a unique `hash()` to check if it's a duplicate, in the
open set, or closed set. Also good for debugging!
- Each node has a `reconstructPath()` function that returns a list of
states that led to this node, starting from the root

The old API is removed and replaced with this instead.